+++ Server/utils/queries.js
@@ -1,7 +1,18 @@
 import { gql } from "@apollo/client";
-new GraphQLObjectType({
-  name: "Query",
-  fields: {
-    _dummy: { type: graphql.GraphQLString },
-  },
-});
+export const QUERY_USER = gql`
+  query user($username: String!) {
+    user(username: $username) {
+      _id
+      username
+      email
+    }
+  }
+`;
+export const IN_STOCK = gql`
+  query Query {
+    inStock {
+      name
+      inStock
+    }
+  }
+`;
