--- server/package.json
@@ -1,19 +1,0 @@
-{
-  "name": "server",
-  "version": "1.0.0",
-  "description": "",
-  "main": "server.js",
-  "scripts": {
-    "start": "node server.js",
-    "server": "nodemon server.js --ignore client"
-  },
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "apollo-server-express": "^3.10.3",
-    "express": "^4.18.1",
-    "mongoose": "^6.6.4",
-    "nodemon": "^2.0.20",
-    "dotenv": "^16.0.3"
-  }
-}
